statistics: make ticker and histogram extendible (#136)
The changes are:
- make `StatisticsImpl` as a template
- remove internal stats due to https://github.com/facebook/rocksdb/pull/4714
