+++ db/c.cc
@@ -2635,11 +2635,10 @@ void rocksdb_options_set_memtable_huge_page_size(rocksdb_options_t* opt,
   opt->rep.memtable_huge_page_size = v;
 }
 
-void rocksdb_options_set_doubly_skip_list_rep(rocksdb_options_t *opt) {
+void rocksdb_options_set_doubly_skip_list_rep(rocksdb_options_t* opt) {
   rocksdb::MemTableRepFactory* factory = new rocksdb::DoublySkipListFactory();
   opt->rep.memtable_factory.reset(factory);
 }
-
 
 void rocksdb_options_set_hash_skip_list_rep(
     rocksdb_options_t *opt, size_t bucket_count,
